chore(releases): moving releases/components to releases/tool/components
