+++ packages/sanity/src/core/releases/tool/detail/ReleaseDetail.tsx
@@ -7,10 +7,10 @@ import {type RouterContextValue, useRouter} from 'sanity/router'
 import {Button} from '../../../../ui-components'
 import {Translate, useTranslation} from '../../../i18n'
 import {useBundles} from '../../../store/bundles'
-import {ReleaseMenuButton} from '../../components/ReleaseMenuButton/ReleaseMenuButton'
-import {ReleasePublishAllButton} from '../../components/ReleasePublishAllButton/ReleasePublishAllButton'
 import {releasesLocaleNamespace} from '../../i18n'
 import {type ReleasesRouterState} from '../../types/router'
+import {ReleaseMenuButton} from '../components/ReleaseMenuButton/ReleaseMenuButton'
+import {ReleasePublishAllButton} from '../components/ReleasePublishAllButton/ReleasePublishAllButton'
 import {useReleaseHistory} from './documentTable/useReleaseHistory'
 import {ReleaseReview} from './ReleaseReview'
 import {ReleaseSummary} from './ReleaseSummary'
