+++ packages/sanity/src/core/releases/tool/detail/ReleaseSummary.tsx
@@ -11,9 +11,9 @@ import {UserAvatar} from '../../../components/userAvatar/UserAvatar'
 import {Translate, useTranslation} from '../../../i18n'
 import {type BundleDocument} from '../../../store/bundles/types'
 import {useBundleOperations} from '../../../store/bundles/useBundleOperations'
-import {Chip} from '../../components/Chip'
-import {Table, type TableProps} from '../../components/Table/Table'
 import {releasesLocaleNamespace} from '../../i18n'
+import {Chip} from '../components/Chip'
+import {Table, type TableProps} from '../components/Table/Table'
 import {DocumentActions} from './documentTable/DocumentActions'
 import {getDocumentTableColumnDefs} from './documentTable/DocumentTableColumnDefs'
 import {type DocumentHistory} from './documentTable/useReleaseHistory'
