+++ packages/sanity/src/core/releases/tool/overview/ReleasesOverview.tsx
@@ -7,11 +7,11 @@ import {Button, Button as StudioButton} from '../../../../ui-components'
 import {ReleaseDetailsDialog} from '../../../bundles/components/dialog/ReleaseDetailsDialog'
 import {useTranslation} from '../../../i18n'
 import {type BundleDocument, useBundles} from '../../../store'
-import {ReleaseMenuButton} from '../../components/ReleaseMenuButton/ReleaseMenuButton'
-import {Table, type TableProps} from '../../components/Table/Table'
-import {type TableSort} from '../../components/Table/TableProvider'
 import {releasesLocaleNamespace} from '../../i18n'
 import {containsBundles} from '../../types/bundle'
+import {ReleaseMenuButton} from '../components/ReleaseMenuButton/ReleaseMenuButton'
+import {Table, type TableProps} from '../components/Table/Table'
+import {type TableSort} from '../components/Table/TableProvider'
 import {type BundlesMetadata, useBundlesMetadata} from '../useBundlesMetadata'
 import {releasesOverviewColumnDefs} from './ReleasesOverviewColumnDefs'
